# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mafia World Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /players/me:
    get:
      operationId: getMyProfile
      tags: [players]
      summary: Get current player profile
      responses:
        "200":
          description: Player profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
    patch:
      operationId: updateMyProfile
      tags: [players]
      summary: Update current player profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerUpdate"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players/me/restart:
    post:
      operationId: restartAfterDeath
      tags: [players]
      summary: Reset the current player after a permanent death
      responses:
        "200":
          description: Restart successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestartResponse"
        "400":
          description: Player is not dead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /players/me/anti-spy:
    post:
      operationId: toggleAntiSpy
      tags: [players]
      summary: Toggle anti-spy mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AntiSpyToggle"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List all players
      parameters:
        - in: query
          name: cityId
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerListResponse"
  /players/{playerId}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get a player by ID
      parameters:
        - in: path
          name: playerId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"

  # --- PROPERTIES ---
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Player:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        username:
          type: string
        level:
          type: integer
        xp:
          type: integer
        money:
          type: integer
        attackPower:
          type: integer
        defensePower:
          type: integer
        killCount:
          type: integer
        deathCount:
          type: integer
        cityId:
          type: integer
        cityName:
          type: string
        gangId:
          type: ["integer", "null"]
        gangName:
          type: ["string", "null"]
        gangRank:
          type: ["string", "null"]
        antiSpyEnabled:
          type: boolean
        isInPrison:
          type: boolean
        prisonReleaseAt:
          type: ["string", "null"]
        isTraveling:
          type: boolean
        travelArrivalAt:
          type: ["string", "null"]
        createdAt:
          type: string
        currentRank:
          type: integer
        rankNameEn:
          type: string
        rankNameAr:
          type: string
        rankColor:
          type: string
        isPermanentlyDead:
          type: boolean
        diedAt:
          type: ["string", "null"]
        killedByPlayerId:
          type: ["integer", "null"]
        killedByUsername:
          type: ["string", "null"]
        deathCause:
          type: ["string", "null"]
      required:
        - id
        - clerkId
        - username
        - level
        - xp
        - money
        - attackPower
        - defensePower
        - killCount
        - deathCount
        - cityId
        - cityName
        - gangId
        - gangName
        - gangRank
        - antiSpyEnabled
        - isInPrison
        - prisonReleaseAt
        - isTraveling
        - travelArrivalAt
        - createdAt
        - currentRank
        - rankNameEn
        - rankNameAr
        - rankColor
        - isPermanentlyDead
        - diedAt
        - killedByPlayerId
        - deathCause
    RestartResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
      required:
        - ok
        - message
    PlayerUpdate:
      type: object
      properties:
        username:
          type: string
    AntiSpyToggle:
      type: object
      properties:
        enabled:
          type: boolean
      required:
        - enabled
    PlayerListResponse:
      type: object
      properties:
        players:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - players
        - total
        - page
        - limit
...