# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mafia World Game API
 * OpenAPI spec version: 0.1.0
 */

export interface Player {
  id: number;
  clerkId: string;
  username: string;
  level: number;
  xp: number;
  money: number;
  attackPower: number;
  defensePower: number;
  killCount: number;
  deathCount: number;
  cityId: number;
  cityName: string;
  /** @nullable */
  gangId: number | null;
  /** @nullable */
  gangName: string | null;
  /** @nullable */
  gangRank: string | null;
  antiSpyEnabled: boolean;
  isInPrison: boolean;
  /** @nullable */
  prisonReleaseAt: string | null;
  isTraveling: boolean;
  /** @nullable */
  travelArrivalAt: string | null;
  createdAt: string;
  currentRank: number;
  rankNameEn: string;
  rankNameAr: string;
  rankColor: string;
  isPermanentlyDead: boolean;
  /** @nullable */
  diedAt: string | null;
  /** @nullable */
  killedByPlayerId: number | null;
  /** @nullable */
  killedByUsername?: string | null;
  /** @nullable */
  deathCause: string | null;
}