# lib/api-zod/src/generated/types/restartResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mafia World Game API
 * OpenAPI spec version: 0.1.0
 */

export interface RestartResponse {
  ok: boolean;
  message: string;
}